requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
pandas==0.24.2
pyparsing==2.3.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
statsmodels==0.9.0
